refactor(core): split episode/retention into submodules
Split large retention.rs file into module directory structure:
- mod.rs: Public API and exports
- policy.rs: Retention policy implementation
- types.rs: CleanupResult and constants
- tests.rs: Unit tests

File size compliance: Each file now <500 LOC.
